long id

+++ src/com/ViasApp/movimiento/Semaforo.scala
@@ -4,7 +4,7 @@ import com.ViasApp.inmovil.Velocidad
 
 
 //viaAsociada no sè si guardarla acà
-class Semaforo(val id:Int,var estado:Int, val tv:Double) {
+class Semaforo(val id:Long,var estado:Int, val tv:Double) {
   //Estado 0:Verde
   //Estado 1:Amarillo
   //estado 2:Rojo
